## Description


https://github.com/ente-io/ente/assets/254676/7ec2f82a-d0d1-4dbe-8836-87d1ffd793f6


## Tests

- [x] Hide Shared Item on Home Page If user has copied it.
- [x] Keep the fav state on the UI
- [x] Client side hash based dedupe works for copied items. Items are
linked to existing item instead of uploading duplicate entries

+++ mobile/lib/services/collections_service.dart
@@ -28,9 +28,11 @@ import 'package:photos/models/collection/collection.dart';
 import 'package:photos/models/collection/collection_file_item.dart';
 import 'package:photos/models/collection/collection_items.dart';
 import 'package:photos/models/file/file.dart';
+import "package:photos/models/files_split.dart";
 import "package:photos/models/metadata/collection_magic.dart";
 import 'package:photos/services/app_lifecycle_service.dart';
 import "package:photos/services/favorites_service.dart";
+import "package:photos/services/feature_flag_service.dart";
 import 'package:photos/services/file_magic_service.dart';
 import 'package:photos/services/local_sync_service.dart';
 import 'package:photos/services/remote_sync_service.dart';
@@ -1148,11 +1150,56 @@ class CollectionsService {
     return collection;
   }
 
-  Future<void> addToCollection(int collectionID, List<EnteFile> files) async {
-    final containsUploadedFile = files.firstWhereOrNull(
-          (element) => element.uploadedFileID != null,
-        ) !=
-        null;
+  Future<void> addOrCopyToCollection(
+    int dstCollectionID,
+    List<EnteFile> files,
+  ) async {
+    final splitResult = FilesSplit.split(files, _config.getUserID()!);
+    if (splitResult.pendingUploads.isNotEmpty) {
+      throw ArgumentError('File should be already uploaded');
+    }
+    if (splitResult.ownedByCurrentUser.isNotEmpty) {
+      await _addToCollection(dstCollectionID, splitResult.ownedByCurrentUser);
+    }
+    if (splitResult.ownedByOtherUsers.isNotEmpty) {
+      if (!FeatureFlagService.instance.isInternalUserOrDebugBuild()) {
+        throw ArgumentError('Cannot add files owned by other users');
+      }
+      late final List<EnteFile> filesToCopy;
+      late final List<EnteFile> filesToAdd;
+      (filesToAdd, filesToCopy) = (await _splitFilesToAddAndCopy(
+        splitResult.ownedByOtherUsers,
+      ));
+
+      if (filesToAdd.isNotEmpty) {
+        _logger.info(
+          "found existing ${filesToAdd.length} files with same hash, adding symlinks",
+        );
+        await _addToCollection(dstCollectionID, filesToAdd);
+      }
+      // group files by collectionID
+      final Map<int, List<EnteFile>> filesByCollection = {};
+      for (final file in filesToCopy) {
+        if (filesByCollection.containsKey(file.collectionID!)) {
+          filesByCollection[file.collectionID!]!.add(file.copyWith());
+        } else {
+          filesByCollection[file.collectionID!] = [file.copyWith()];
+        }
+      }
+      for (final entry in filesByCollection.entries) {
+        final srcCollectionID = entry.key;
+        final files = entry.value;
+        await _copyToCollection(
+          files,
+          dstCollectionID: dstCollectionID,
+          srcCollectionID: srcCollectionID,
+        );
+      }
+    }
+  }
+
+  Future<void> _addToCollection(int collectionID, List<EnteFile> files) async {
+    final containsUploadedFile = files.any((e) => e.isUploaded);
     if (containsUploadedFile) {
       final existingFileIDsInCollection =
           await FilesDB.instance.getUploadedFileIDs(collectionID);
@@ -1165,6 +1212,13 @@ class CollectionsService {
     if (files.isEmpty || !containsUploadedFile) {
       _logger.info("nothing to add to the collection");
       return;
+    }
+    final anyFileOwnedByOther =
+        files.any((e) => e.ownerID != null && e.ownerID != _config.getUserID());
+    if (anyFileOwnedByOther) {
+      throw ArgumentError(
+        'Cannot add files owned by other users, they should be copied',
+      );
     }
 
     final params = <String, dynamic>{};
@@ -1259,6 +1313,126 @@ class CollectionsService {
       } catch (e) {
         _logger.warning('failed to add files to collection', e);
         rethrow;
+      }
+    }
+  }
+
+  Future<void> _copyToCollection(
+    List<EnteFile> files, {
+    required int dstCollectionID,
+    required int srcCollectionID,
+  }) async {
+    _validateCopyInput(dstCollectionID, srcCollectionID, files);
+    final batchedFiles = files.chunks(batchSize);
+    final params = <String, dynamic>{};
+    params["dstCollectionID"] = dstCollectionID;
+    params["srcCollectionID"] = srcCollectionID;
+    for (final batch in batchedFiles) {
+      params["files"] = [];
+      for (final batchFile in batch) {
+        final fileKey = getFileKey(batchFile);
+        _logger.info(
+          "srcCollection : $srcCollectionID  file: ${batchFile.uploadedFileID}  key: ${CryptoUtil.bin2base64(fileKey)} ",
+        );
+        final encryptedKeyData =
+            CryptoUtil.encryptSync(fileKey, getCollectionKey(dstCollectionID));
+        batchFile.encryptedKey =
+            CryptoUtil.bin2base64(encryptedKeyData.encryptedData!);
+        batchFile.keyDecryptionNonce =
+            CryptoUtil.bin2base64(encryptedKeyData.nonce!);
+        params["files"].add(
+          CollectionFileItem(
+            batchFile.uploadedFileID!,
+            batchFile.encryptedKey!,
+            batchFile.keyDecryptionNonce!,
+          ).toMap(),
+        );
+      }
+
+      try {
+        final res = await _enteDio.post(
+          "/files/copy",
+          data: params,
+        );
+        final oldToCopiedFileIDMap = Map<int, int>.from(
+          (res.data["oldToNewFileIDMap"] as Map<String, dynamic>).map(
+            (key, value) => MapEntry(int.parse(key), value as int),
+          ),
+        );
+        for (final file in batch) {
+          final int uploadIDForOriginalFIle = file.uploadedFileID!;
+          if (oldToCopiedFileIDMap.containsKey(uploadIDForOriginalFIle)) {
+            file.generatedID = null;
+            file.collectionID = dstCollectionID;
+            file.uploadedFileID = oldToCopiedFileIDMap[uploadIDForOriginalFIle];
+            file.ownerID = _config.getUserID();
+            oldToCopiedFileIDMap.remove(uploadIDForOriginalFIle);
+          } else {
+            throw Exception("Failed to copy file ${file.uploadedFileID}");
+          }
+        }
+        if (oldToCopiedFileIDMap.isNotEmpty) {
+          throw Exception(
+            "Failed to map following uploadKey ${oldToCopiedFileIDMap.keys}",
+          );
+        }
+        await _filesDB.insertMultiple(batch);
+        Bus.instance
+            .fire(CollectionUpdatedEvent(dstCollectionID, batch, "copiedTo"));
+      } catch (e) {
+        rethrow;
+      }
+    }
+  }
+
+  Future<(List<EnteFile>, List<EnteFile>)> _splitFilesToAddAndCopy(
+    List<EnteFile> othersFile,
+  ) async {
+    final hashToUserFile =
+        await _filesDB.getUserOwnedFilesWithSameHashForGivenListOfFiles(
+      othersFile,
+      _config.getUserID()!,
+    );
+    final List<EnteFile> filesToCopy = [];
+    final List<EnteFile> filesToAdd = [];
+    for (final EnteFile file in othersFile) {
+      if (hashToUserFile.containsKey(file.hash ?? '')) {
+        final userFile = hashToUserFile[file.hash]!;
+        if (userFile.fileType == file.fileType) {
+          filesToAdd.add(userFile);
+        } else {
+          filesToCopy.add(file);
+        }
+      } else {
+        filesToCopy.add(file);
+      }
+    }
+    return (filesToAdd, filesToCopy);
+  }
+
+  void _validateCopyInput(
+    int destCollectionID,
+    int srcCollectionID,
+    List<EnteFile> files,
+  ) {
+    final dstCollection = _collectionIDToCollections[destCollectionID];
+    final srcCollection = _collectionIDToCollections[srcCollectionID];
+    if (dstCollection == null || !dstCollection.isOwner(_config.getUserID()!)) {
+      throw ArgumentError(
+        'Destination collection not found ${dstCollection == null} or not owned by user ',
+      );
+    }
+    if (srcCollection == null) {
+      throw ArgumentError('Source collection not found');
+    }
+    // verify that all fileIds belong to srcCollection and isn't owned by current user
+    for (final f in files) {
+      if (f.collectionID != srcCollectionID ||
+          f.ownerID == _config.getUserID()) {
+        _logger.warning(
+          'file $f does not belong to srcCollection $srcCollection or is owned by current user ${f.ownerID}',
+        );
+        throw ArgumentError('');
       }
     }
   }
@@ -1481,10 +1655,13 @@ class CollectionsService {
       for (final file in batch) {
         params["fileIDs"].add(file.uploadedFileID);
       }
-      await _enteDio.post(
+      final resp = await _enteDio.post(
         "/collections/v3/remove-files",
         data: params,
       );
+      if (resp.statusCode != 200) {
+        throw Exception("Failed to remove files from collection");
+      }
 
       await _filesDB.removeFromCollection(collectionID, params["fileIDs"]);
       Bus.instance
